role dropdown selecting
